+++ src/model/playlist.rs
@@ -6,9 +6,8 @@ use std::collections::HashMap;
 
 use super::image::Image;
 use super::page::Page;
-use super::track::FullTrack;
 use super::user::PublicUser;
-use crate::model::{Followers, Type};
+use crate::model::{Followers, PlayableItem, Type};
 
 /// Playlist result object
 ///
@@ -77,7 +76,7 @@ pub struct PlaylistItem {
     pub added_at: Option<DateTime<Utc>>,
     pub added_by: Option<PublicUser>,
     pub is_local: bool,
-    pub track: Option<FullTrack>,
+    pub track: Option<PlayableItem>,
 }
 /// Featured playlists object
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/#endpoint-get-featured-playlists)